+++ secure-api-gateway-ob-uk-rcs-consent-store/secure-api-gateway-ob-uk-rcs-consent-store-api/src/main/java/com/forgerock/sapi/gateway/rcs/consent/store/api/payment/domesticscheduled/BaseDomesticScheduledPaymentConsentApiController.java
@@ -1,6 +1,22 @@
+/*
+ * Copyright © 2020-2024 ForgeRock AS (obst@forgerock.com)
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
 package com.forgerock.sapi.gateway.rcs.consent.store.api.payment.domesticscheduled;
 
-import java.util.Objects;
+import static java.util.Objects.requireNonNull;
+
 import java.util.function.Supplier;
 
 import org.joda.time.DateTime;
@@ -21,20 +37,28 @@ import com.forgerock.sapi.gateway.uk.common.shared.api.meta.obie.OBVersion;
 
 import uk.org.openbanking.datamodel.v3.payment.OBPaymentConsentStatus;
 
+/**
+ * Base implementation of {@link DomesticScheduledPaymentConsentApi} containing the logic that is common across versions.
+ * The {@link OBVersion} can be plugged in via the constructor.
+ * <p>
+ * Version specific controller implementations should extend this class and specify the following annotations:
+ * {@link org.springframework.stereotype.Controller}, {@link org.springframework.web.bind.annotation.RequestMapping}
+ * and {@link io.swagger.annotations.Api}.
+ */
 public class BaseDomesticScheduledPaymentConsentApiController implements DomesticScheduledPaymentConsentApi {
 
     private final Logger logger = LoggerFactory.getLogger(getClass());
 
-    protected final DomesticScheduledPaymentConsentService consentService;
-    protected final Supplier<DateTime> idempotencyKeyExpirationSupplier;
-    protected final OBVersion obVersion;
+    private final DomesticScheduledPaymentConsentService consentService;
+    private final Supplier<DateTime> idempotencyKeyExpirationSupplier;
+    private final OBVersion obVersion;
 
     public BaseDomesticScheduledPaymentConsentApiController(DomesticScheduledPaymentConsentService consentService,
                                                             Supplier<DateTime> idempotencyKeyExpirationSupplier,
                                                             OBVersion obVersion) {
-        this.consentService = Objects.requireNonNull(consentService, "consentService must be provided");
-        this.idempotencyKeyExpirationSupplier = Objects.requireNonNull(idempotencyKeyExpirationSupplier, "idempotencyKeyExpirationSupplier must be provided");
-        this.obVersion = Objects.requireNonNull(obVersion, "obVersion must be provided");
+        this.consentService = requireNonNull(consentService, "consentService must be provided");
+        this.idempotencyKeyExpirationSupplier = requireNonNull(idempotencyKeyExpirationSupplier, "idempotencyKeyExpirationSupplier must be provided");
+        this.obVersion = requireNonNull(obVersion, "obVersion must be provided");
     }
 
     @Override
